Fix value used for test to pass without GMP and with windows
* So long == 32 bits and long long == 64 bits

+++ pythran/tests/euler/euler10.py
@@ -3,7 +3,7 @@
 def solve(max):
     '''
     The sum of the primes below 10 is 2 + 3 + 5 + 7 = 17.
-    
+
     Find the sum of all the primes below two million.
     '''
 
@@ -16,4 +16,4 @@ def solve(max):
     for x in xrange(2, int(len(sieve) ** 0.5) + 1):
         if sieve[x]: mark(sieve, x)
 
-    return sum(i for i in xrange(2, len(sieve)) if sieve[i])
+    return sum(long(i) for i in xrange(2, len(sieve)) if sieve[i])
